+++ table_gen.cpp
@@ -2,8 +2,8 @@
 #include <math.h>
 #include <vector>
 #include <bitset>
-#include "headers.hpp"
 #include "move_tables.hpp"
+#include <cstdint>
 
 /* Bitboard configuration
     29    30    31    32
@@ -17,30 +17,30 @@
 */
 
 //prints a binary representation of a number
-void print_binary(long long num){
+void print_binary(uint32_t num){
     std::cout << std::bitset<32>(num) << std::endl;
 }
 
 //converts a square index (according to the chart above) to its bitboard representation
-long long square_to_binary(const int square){
-    long long n = 0b1;
+uint32_t square_to_binary(const int square){
+    uint32_t n = 0b1;
     n = n << (square - 1);
     return n;
 }
 
 //gets the square index of a binary number
 //NOTE: THERE CAN ONLY BE ONE 1 BIT
-int binary_to_square(const long long binary){
+int binary_to_square(const uint32_t binary){
     return log2(binary) + 1;
 }
 
 //gets the squares of all 1 bits in a bitboard
-std::vector<int> serialize_bb(long long bb){
+std::vector<int> serialize_bb(uint32_t bb){
     std::vector<int> squares;
     if(bb == 0){return squares;}
 
     while (bb){
-        long long ls1b = bb & -bb; // isolate LS1B
+        uint32_t ls1b = bb & -bb; // isolate LS1B
         squares.push_back(binary_to_square(ls1b)); //add squares to vector
         bb &= bb-1; // reset LS1B
     }
@@ -49,7 +49,7 @@ std::vector<int> serialize_bb(long long bb){
 
 //generates moves for any given square
 //NOTE: currently configured for black takes
-long long gen_mask(const double square){
+uint32_t gen_mask(const double square){
     int row = ceil(square/4);
     int west = -9;
     int east = -7;
@@ -65,7 +65,7 @@ long long gen_mask(const double square){
     //     east = -4;
     // }
 
-    long long bb = 0b0;
+    uint32_t bb = 0b0;
     target1 = square + west;
     target2 = square + east;
 
@@ -80,7 +80,7 @@ long long gen_mask(const double square){
 
 
 int main(){
-    long long moves[33];
+    uint32_t moves[33];
     for (int i = 1; i <= 32; i++){
         moves[i] = red_takes[i] | black_takes[i];
     }
